[#57] search 모듈 해체 - 각 도메인 모듈로 검색 로직 이전 및 엔드포인트 재설계
