+++ src/main/java/com/backend/allreva/module/concert/concert/application/dto/SortDirection.java
@@ -1,0 +1,7 @@
+package com.backend.allreva.module.concert.concert.application.dto;
+
+public enum SortDirection {
+    DATE,
+    VIEWS,
+    SCORE
+}
