+++ src/main/java/com/backend/allreva/module/recruitment/rent/application/dto/RentThumbnail.java
@@ -1,4 +1,4 @@
-package com.backend.allreva.module.search.application.dto;
+package com.backend.allreva.module.recruitment.rent.application.dto;
 
 import java.time.LocalDate;
