+++ src/main/java/com/backend/allreva/module/recruitment/rent/application/port/RentSearchRepository.java
@@ -1,7 +1,7 @@
-package com.backend.allreva.module.search.application.port;
+package com.backend.allreva.module.recruitment.rent.application.port;
 
 import com.backend.allreva.common.pagination.SliceResponse;
-import com.backend.allreva.module.search.application.dto.RentThumbnail;
+import com.backend.allreva.module.recruitment.rent.application.dto.RentThumbnail;
 import java.util.List;
 
 public interface RentSearchRepository {
