--- src/main/java/com/backend/allreva/module/search/application/dto/SortDirection.java
@@ -1,7 +1,0 @@
-package com.backend.allreva.module.search.application.dto;
-
-public enum SortDirection {
-    DATE,
-    VIEWS,
-    SCORE
-}
