--- src/main/java/com/backend/allreva/module/search/application/port/PlaceSearchRepository.java
@@ -1,8 +1,0 @@
-package com.backend.allreva.module.search.application.port;
-
-import com.backend.allreva.module.concert.place.application.dto.ConcertHallMainResponse;
-
-public interface PlaceSearchRepository {
-
-    ConcertHallMainResponse searchMain(String address, int minSeatSize, String cursorId, int pageSize);
-}
