--- src/test/java/com/backend/allreva/module/search/integration/PlaceSearchServiceTest.java
@@ -1,145 +1,0 @@
-package com.backend.allreva.module.search.integration;
-
-import static org.assertj.core.api.SoftAssertions.assertSoftly;
-import static org.instancio.Select.field;
-
-import com.backend.allreva.module.concert.place.application.dto.ConcertHallMainResponse;
-import com.backend.allreva.module.concert.place.domain.ConcertHall;
-import com.backend.allreva.module.concert.place.domain.ConcertHallRepository;
-import com.backend.allreva.module.concert.place.domain.value.Location;
-import com.backend.allreva.module.concert.place.fixture.ConcertHallFixture;
-import com.backend.allreva.module.search.application.PlaceSearchService;
-import com.backend.allreva.support.IntegrationTestSupport;
-import java.util.Objects;
-import lombok.extern.slf4j.Slf4j;
-import org.instancio.Instancio;
-import org.junit.jupiter.api.AfterEach;
-import org.junit.jupiter.api.Disabled;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Nested;
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Qualifier;
-import org.springframework.cache.CacheManager;
-
-@Slf4j
-@SuppressWarnings("NonAsciiCharacters")
-@DisplayName("Place 검색 통합 테스트")
-class PlaceSearchServiceTest extends IntegrationTestSupport {
-
-    @Autowired
-    PlaceSearchService placeSearchService;
-
-    @Autowired
-    ConcertHallRepository concertHallRepository;
-
-    @Autowired
-    @Qualifier("placeMainCacheManager")
-    CacheManager cacheManager;
-
-    @AfterEach
-    void tearDown() {
-        concertHallRepository.deleteAll();
-        Objects.requireNonNull(cacheManager.getCache("placeMain")).clear();
-    }
-
-    @Nested
-    @DisplayName("메인 공연장 검색")
-    class Describe_메인_공연장_검색 {
-
-        @Nested
-        @DisplayName("지역으로 검색할 때")
-        class Context_지역으로_검색 {
-
-            @Test
-            @DisplayName("해당 지역의 공연장이 반환된다")
-            void 해당_지역의_공연장이_반환된다() {
-                // given
-                concertHallRepository.save(Instancio.of(ConcertHallFixture.concertHallModel())
-                        .set(field(Location.class, "address"), "서울특별시 종로구")
-                        .set(field(ConcertHall.class, "seatScale"), 5000)
-                        .create());
-                String address = "서울";
-                int seatScale = 0;
-                int size = 10;
-
-                // when
-                ConcertHallMainResponse result = placeSearchService.searchMainPlaces(address, seatScale, null, size);
-
-                // then
-                assertSoftly(softly -> {
-                    softly.assertThat(result).isNotNull();
-                    softly.assertThat(result.concertHallThumbnails()).isNotEmpty();
-                });
-            }
-        }
-
-        @Nested
-        @DisplayName("좌석 규모로 필터링할 때")
-        class Context_좌석_규모로_필터링 {
-
-            @Test
-            @DisplayName("지정한 좌석 규모 이상의 공연장만 반환된다")
-            void 지정한_좌석_규모_이상의_공연장만_반환된다() {
-                // given
-                concertHallRepository.save(Instancio.of(ConcertHallFixture.concertHallModel())
-                        .set(field(Location.class, "address"), "서울특별시 종로구")
-                        .set(field(ConcertHall.class, "seatScale"), 5000)
-                        .create());
-                String address = "";
-                int seatScale = 2000;
-                int size = 10;
-
-                // when
-                ConcertHallMainResponse result = placeSearchService.searchMainPlaces(address, seatScale, null, size);
-
-                // then
-                assertSoftly(softly -> {
-                    softly.assertThat(result).isNotNull();
-                    softly.assertThat(result.concertHallThumbnails()).isNotEmpty();
-                    softly.assertThat(result.concertHallThumbnails()).allMatch(hall -> hall.seatScale() >= seatScale);
-                });
-            }
-        }
-
-        @Nested
-        @DisplayName("페이지네이션을 사용한 조회 시")
-        class Context_페이지네이션_조회 {
-
-            // TODO: 작업 중인 테스트
-            @Disabled("작업 중인 테스트")
-            @Test
-            @DisplayName("nextCursorId를 사용하여 다음 페이지를 조회할 수 있다")
-            void nextCursorId를_사용하여_다음_페이지를_조회할_수_있다() {
-                // given - 각각 다른 ID로 5개 저장
-                for (int i = 1; i <= 5; i++) {
-                    concertHallRepository.save(Instancio.of(ConcertHallFixture.concertHallModel())
-                            .set(field(ConcertHall.class, "hallCode"), String.format("hall-%03d", i))
-                            .create());
-                }
-                String address = "";
-                int seatScale = 0;
-                int pageSize = 2;
-
-                // when - 첫 페이지 조회
-                ConcertHallMainResponse page1 = placeSearchService.searchMainPlaces(address, seatScale, null, pageSize);
-
-                // then
-                assertSoftly(softly -> {
-                    softly.assertThat(page1.concertHallThumbnails()).hasSize(2);
-                    softly.assertThat(page1.nextCursorId()).isNotNull();
-                });
-
-                // when - 두 번째 페이지 조회
-                ConcertHallMainResponse page2 =
-                        placeSearchService.searchMainPlaces(address, seatScale, page1.nextCursorId(), pageSize);
-
-                // then
-                assertSoftly(softly -> {
-                    softly.assertThat(page2.concertHallThumbnails()).hasSize(2);
-                    softly.assertThat(page2.concertHallThumbnails()).isNotEqualTo(page1.concertHallThumbnails());
-                });
-            }
-        }
-    }
-}
